webapp/requirements.txt
backports.zoneinfo
crispy-bootstrap5
cryptography
django-allauth
django-freeradius
django-rest-auth
docopt
gunicorn
markdown
pip-chill
psycopg2-binary==2.8.6
pydantic
typing
yarg
WeasyPrint
django-environ
whitenoise
passlib


